#73 - round 1 cleaning up PR
