+++ src/builderForm/build_parameter.ts
@@ -1,5 +1,4 @@
-import { Parameter } from 'types/command_types'
-import { Dictionary } from 'types/custom_types'
+import { Dictionary, Parameter } from 'types/custom_types'
 
 function getTypeAndFormat(parameter: Parameter) {
   let parameterType: string | string[] = parameter.type.toLowerCase()
