+++ src/components/command_view_form.tsx
@@ -15,8 +15,7 @@ import { useState } from 'react'
 import ReactJson from 'react-json-view'
 import { Link as RouterLink, Navigate } from 'react-router-dom'
 import CacheService from 'services/cache_service'
-import { Command } from 'types/command_types'
-import { Dictionary } from 'types/custom_types'
+import { Command, Dictionary } from 'types/custom_types'
 
 interface CommandViewFormProps {
   schema: Dictionary
