+++ src/hooks/useCommands.tsx
@@ -10,7 +10,7 @@ import {
   useState,
 } from 'react'
 import { useParams } from 'react-router-dom'
-import { Command, CommandRow } from 'types/command_types'
+import { Command, CommandRow } from 'types/custom_types'
 import { System } from 'types/custom_types'
 
 export const useCommands = () => {
