+++ src/pages/CommandIndex/commandIndexHelpers.tsx
@@ -1,7 +1,6 @@
 import { MakeItHappenButton } from 'pages/CommandIndex'
 import { filterSystems } from 'services/system.service'
-import { Command } from 'types/command_types'
-import { System } from 'types/custom_types'
+import { Command, System } from 'types/custom_types'
 
 const formatCommands = (commands: Command[]) => {
   const formattedData: (string | JSX.Element)[][] = []
