+++ src/pages/RequestView/RequestView.tsx
@@ -78,32 +78,38 @@ const RequestView = () => {
 
   const pourItAgainClick = () => {
     if (request) {
-      CacheService.pushQueue(request, 'lastKnownPourItAgainRequest')
+      CacheService.pushQueue(request, `lastKnownPourItAgainRequest`)
     }
   }
 
   return (
     <Box>
-      <Typography style={{ flex: 1, float: 'right' }}>
-        <Button
-          component={RouterLink}
-          to={[
-            '/systems',
-            namespace,
-            system,
-            systemVersion,
-            'commands',
-            command,
-          ].join('/')}
-          variant="contained"
-          color="primary"
-          onAuxClick={pourItAgainClick}
-          onClick={pourItAgainClick}
-        >
-          Pour it Again
-        </Button>
-      </Typography>
-      <PageHeader title="Request View" description={String(id)} />
+      <Grid justifyContent="space-between" container>
+        <Grid item>
+          <PageHeader title="Request View" description={String(id)} />
+        </Grid>
+        <Grid item>
+          <Typography style={{ flex: 1 }}>
+            <Button
+              component={RouterLink}
+              to={[
+                '/systems',
+                namespace,
+                system,
+                systemVersion,
+                'commands',
+                command,
+              ].join('/')}
+              variant="contained"
+              color="primary"
+              onAuxClick={pourItAgainClick}
+              onClick={pourItAgainClick}
+            >
+              Pour it Again
+            </Button>
+          </Typography>
+        </Grid>
+      </Grid>
       <Divider />
       {request ? (
         <>
