+++ src/services/system.service.ts
@@ -1,7 +1,6 @@
 import { AxiosRequestConfig, AxiosResponse } from 'axios'
 import { useMyAxios } from 'hooks/useMyAxios'
-import { Command } from 'types/command_types'
-import { SuccessCallback, System } from 'types/custom_types'
+import { Command, SuccessCallback, System } from 'types/custom_types'
 
 const getSystem = (
   systems: System[],
