--- src/types/props_types.ts
@@ -1,8 +1,0 @@
-export interface modalProps {
-  header: string
-  open: boolean
-  content: JSX.Element
-  onClose(): void
-  onCancel(): void
-  onSubmit(): void
-}
